Add initial draft of scripts for post-release and CESM update
Post-release workflow gets the develop branches of every EarthWorksOrg
external ready for the next development cycle. This includes merging the
latest release tag into the ew-develop branches, creating a new tag
ending in 000 for each ext, updating EarthWorks/Externals.cfg with new
tags, tagging that in EarthWorks, and then maybe pushing all updated
branches & tags.

CESM update uses a CESM tag to fetch it's Externals.cfg and for each
external we've forked into EarthWorksOrg the script will find the tag
from CESM_Externals.cfg, create a new branch for the update, merge the
update tag into ew-develop, create a tag for that commit, and maybe push
the results. The script will then modify the EW_Externals.cfg with the
new tags of EarthWorksOrgs that were created and the updated tags from
CESM_Externals.cfg for the externals we still share with CESM.