Switch CTSM (clm) back to the upstream version
Specifically edit the clm entry to use the ESCOMP/CTSM url and
ctsm5.3.002 tag. This matches with what is used in ESCOMP/CESM in the
cesm3_0_beta03 tag.

This prevents EarthWorks from having issues reading and using files
generated by the updated CTSM and decreases the drift between EarthWorks
and CESM.